default results are fetched when navigating to /searchResults
