+++ src/app/app.component.ts
@@ -1,34 +1,9 @@
-import { Component, OnInit } from '@angular/core';
-import { ApiService } from './services/_http/http-service.service';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
-  providers: [ ApiService ]
+  styleUrls: ['./app.component.css']
 })
-export class AppComponent implements OnInit {
 
-  public _offers;
-  private _busy: boolean;
-  constructor(private _apiService: ApiService) {
-    this._busy = true;
-  }
-
-  public getDeals() {
-    this._apiService.getDeals().subscribe(
-      data => { this._offers = data; },
-      err => console.error(err),
-      () => {
-        console.log( this._offers.deals[2].images, 'deals loading complete' ),
-        this._busy = false;
-      }
-    );
-
-  }
-
-  ngOnInit() {
-    // this.getDeals();
-  }
-
-}
+export class AppComponent { }
