+++ src/app/modules/search-results/search-results.component.ts
@@ -1,26 +1,49 @@
-import { Component, OnDestroy } from '@angular/core';
+import {Component, OnDestroy, OnInit } from '@angular/core';
 import { Router, NavigationEnd } from '@angular/router';
+import { ApiService } from '../../services/_http/http-service.service';
 
 @Component({
-    selector: 'app-search-results-main',
-    templateUrl: 'search-results.template.html',
-    styleUrls: ['./search-results.style.css']
+  selector: 'app-search-results-main',
+  templateUrl: 'search-results.template.html',
+  styleUrls: ['./search-results.style.css'],
+  providers: [ ApiService ]
 })
 
-export class SearchResultsIndexComponent implements OnDestroy {
-    isParentActive: boolean;
-    routerSubscription: any;
-    componentsHeader = 'This is the search results page!!!!';
+export class SearchResultsIndexComponent implements OnInit, OnDestroy {
+  isParentActive: boolean;
+  routerSubscription: any;
+  componentsHeader = 'This is the search results page!!!!';
+  public _offers;
+  public _busy: boolean;
 
-    constructor(private router: Router) {
-      this.routerSubscription = this.router.events.subscribe(event => {
-        if (event instanceof NavigationEnd) {
-          this.isParentActive = event.url === '/searchResults';
-          }
-      });
-    }
+  constructor(private _router: Router, private _apiService: ApiService) {
+    // Router Subscription:
+    this.routerSubscription = this._router.events.subscribe(event => {
+      if (event instanceof NavigationEnd) {
+        this.isParentActive = event.url === '/searchResults';
+        }
+    });
 
-    ngOnDestroy() {
-      this.routerSubscription.unsubscribe();
-    }
+    // Results loading (getDeals):
+    this._busy = true;
+  }
+
+  public getDeals() {
+    this._apiService.getDeals().subscribe(
+      data => { this._offers = data; },
+      err => console.error(err),
+      () => {
+        console.log( this._offers.deals[2].images, 'loading deals complete' ),
+          this._busy = false;
+      }
+    );
+  }
+
+  ngOnInit() {
+    this.getDeals();
+  }
+
+  ngOnDestroy() {
+    this.routerSubscription.unsubscribe();
+  }
 }
