+++ src/app/modules/search-results/search-results.template.html
@@ -9,7 +9,7 @@
     <span *ngIf="!isParentActive"> >
       <a class="[ typography-body-book ]"
          routerLink="/buildingBlocks" routerLinkActive="active">
-        Building Blocks menu
+        Search Results
       </a>
     </span>
   </div>
@@ -26,6 +26,45 @@
     </div>
   </div>
 </div>
+
+<!--<h1> Angular 5 Wowcher Public API demo </h1>-->
+<!--<h2> Deals </h2>-->
+
+<div class="container">
+  <div *ngIf="_busy" class="row">
+    <div class="col-12">
+      <img src="/assets/loading.gif" alt="content loading" />
+    </div>
+  </div>
+
+  <div *ngIf="!_busy" class="row">
+
+    <div class="[ col-4 ]"
+         *ngFor="let offer of _offers?.deals">
+
+      <div class="[ custom__card ]">
+          <span *ngIf="offer?.images.length >= 1">
+            <img class="card-img-top"
+                 src="{{ offer?.images[0].imageUrl }}.jpg" alt="{{ offer?.images[0].alt }}}" />
+          </span>
+        <span *ngIf="offer?.images.length < 1">
+            <img class="card-img-top"
+                 src="http://via.placeholder.com/225x225?text=No%20image%20available" alt="No image available" />
+          </span>
+
+        <div class="card-body">
+          <h5 class="card-title">{{ offer.headline }}</h5>
+          <p class="card-text">Buy now at only £{{ offer.price }}</p>
+          <a href="#" class="btn btn-primary">View deal</a>
+        </div>
+      </div>
+
+    </div>
+
+  </div>
+
+</div>
+
 
 <!--<div class="[ container ][ content__container ]" *ngIf="!isParentActive">-->
   <!--<router-outlet></router-outlet>-->
