+++ src/app/services/_http/http-service.service.ts
@@ -5,7 +5,8 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 const httpOptions = {
   headers: new HttpHeaders({ 'Content-Type' : 'application-json' })
 };
-const url = 'https://public-api.wowcher.co.uk/v1/deal/brighton?pageSize=550';
+// const url = 'https://public-api.wowcher.co.uk/v1/deal/brighton?pageSize=550';
+const baseUrl = 'https://public-api.wowcher.co.uk/v1/deal';
 
 @Injectable()
 export class ApiService {
@@ -13,7 +14,7 @@ export class ApiService {
   constructor(private _http: HttpClient) { }
 
   getDeals() {
-    return this._http.get(url);
+    return this._http.get(baseUrl);
   }
 
 }
